backend/requirements.txt
Django
djangorestframework
dj-rest-auth
django-allauth
django-cors-headers
djangorestframework-simplejwt
requests
cryptography
Pillow
razorpay
fpdf2
ruff
django-axes[ipware]
django-debug-toolbar
pre-commit
pytest
factory-boy
Faker
pyinstrument
celery
python-dotenv
weasyprint
boto3
psycopg
# psycopg2-binary
django-unfold
Faker
django-silk